add purchaseRequest service and controller methods. put and delete

+++ src/main/java/com/cncstock/repository/PurchaseRequestRepository.java
@@ -3,5 +3,5 @@ package com.cncstock.repository;
 import com.cncstock.model.PurchaseRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseRequestRepository extends JpaRepository<PurchaseRequest, Integer> {
+public interface PurchaseRequestRepository extends JpaRepository<PurchaseRequest, Long> {
 }
